# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kommiuniti API",
    "description": "API for Kommiuniti App",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/getUser": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/databyuserid/{userid}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/getAll": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/getcommunitymembers/{communityId}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "communityId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/update": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstname": {
                  "example": "any"
                },
                "lastname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "profile_image": {
                  "example": "any"
                },
                "color": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/sendrequest": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/task/create": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "done": {
                  "example": "any"
                },
                "fk_routine_id": {
                  "example": "any"
                },
                "assignedUser": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/task/delete/{id}": {
      "delete": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/task/gettasksininterval": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "startDate": {
                  "example": "any"
                },
                "endDate": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/task/routine/modify": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "startDate": {
                  "example": "any"
                },
                "interval": {
                  "example": "any"
                },
                "assignedUser": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/task/routine/all": {
      "get": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  }
}